plants and devices context created to resolve the weighted data handling through the pages and websocket issue resolved

+++ code/BackEnd_MobileApp/Green_Tech/src/main/java/com/Green_Tech/Green_Tech/DTO/UserDTO.java
@@ -13,5 +13,5 @@ public class UserDTO {
     String name;
     String email;
     String password;
-    Long phoneNumber;
+    Integer phoneNumber;
 }
